refactor: 핸들러 로직 로컬 hooks로 분리
